Create new models.
Former-commit-id: cb1ed9e5ba8826cf5f5815c39e3313fe7deba53a
Former-commit-id: d377f430fa594d953a96a43e189675ffb76c9b2f

+++ database/migrations/2015_09_02_123653_create_episodes_table.php
@@ -11,7 +11,7 @@ class CreateEpisodesTable extends Migration
         Schema::create('episodes', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('anime_id')->unsigned();
-            $table->integer('number')->unsigned();
+            $table->float('number')->unsigned();
             $table->string('name')->nullable();
             $table->string('slug', 191)->unique()->nullable();
             $table->text('synopsis')->nullable();
